+++ tests/test_format_diff.py
@@ -3,7 +3,7 @@
 ''' Tests for gendiff project.'''
 
 import pytest
-from gendiff.format_diff import format_diff
+from gendiff.formaters.format_diff import format_diff
 
 
 def test_format_diff_jsonlike_formater():
